# contracts/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "VNI IELTS AI",
    "description": "The exam engine, identity, and content APIs. Generated from the running application \u2014 never hand-edited.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/admin/import/packages": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Upload one exam package ZIP; validates and creates a review draft synchronously",
        "operationId": "AdminImportPackage",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "No token, or one this API will not accept."
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/import/packages/{draftId}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Status and findings for one import draft",
        "operationId": "AdminGetImportDraft",
        "parameters": [
          {
            "name": "draftId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "No token, or one this API will not accept."
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/import/packages/{draftId}/warnings/{warningId}/override": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Resolve a warning with a mandatory reason; audited as WarningOverridden",
        "operationId": "AdminOverrideImportWarning",
        "parameters": [
          {
            "name": "draftId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "warningId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideWarningRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "No token, or one this API will not accept."
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/import/packages/{draftId}/approve": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Approve a draft once every warning is resolved and the checklist is complete",
        "operationId": "AdminApproveImportDraft",
        "parameters": [
          {
            "name": "draftId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "No token, or one this API will not accept."
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "OverrideWarningRequest": {
        "required": [
          "reason"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "The access token from /api/v1/auth/login, /auth/refresh or the SSO handoff. Fifteen minutes; the client refreshes ahead of expiry.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}